sandbox/pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.25.0",
    "openai>=1.79.0",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
]
